requirements.txt
python-pptx
opencv-python
alive-progress
pillow